fix(core.cljs): Isolate p5.scale when rendering images
When p5.scale is called, all subsequently drawn shapes in the same draw call will be affected. This means that When rendering two images one after the other, the first image's scale-x or scale-y affects how the second image will be rendered. With the changes in this PR, the scale of the first image will not bleed into the second.